requirements.txt
requests==2.31.0
dataclasses==0.6
python-dateutil==2.8.2